+++ src/main/kotlin/app/revanced/patches/youtube/misc/playercontrols/resource/patch/BottomControlsResourcePatch.kt
@@ -11,13 +11,14 @@ import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patches.shared.mapping.misc.patch.ResourceMappingPatch
 import app.revanced.patches.youtube.misc.playercontrols.annotation.PlayerControlsCompatibility
+import java.io.Closeable
 
 @Name("bottom-controls-resource-patch")
 @Description("Manages the resources for the bottom controls of the YouTube player.")
 @PlayerControlsCompatibility
 @DependsOn([ResourceMappingPatch::class])
 @Version("0.0.1")
-class BottomControlsResourcePatch : ResourcePatch {
+class BottomControlsResourcePatch : ResourcePatch, Closeable {
     override fun execute(context: ResourceContext): PatchResult {
         resourceContext = context
         targetXmlEditor = context.xmlEditor[TARGET_RESOURCE]